Addition of fixed_max() to monitors, CurveView now scales directly with plotted data
fixed_max allows the CurveView to keep a fixed scale for monitors
such as the CPU usage monitor.

CurveView no longer gradually decreases scale of graph over a long
period of time - the scale is tied to the largest value across the
curves. This keeps the scale representative at all times, and prevents
clipping of the curve maxima. This allows for a proper network
bandwidth graph to be implemented

Fixes #1 / https://github.com/OmegaPhil/hardware-monitor-applet/issues/1